refactor: SideBarItem 컴포넌트로 분리하여 사용 #5
